+++ app.rb
@@ -1,29 +1,28 @@
 require './display'
-require './userInteraction'
+require './user_interaction'
 # app class
 class App
   attr_reader :create_classes
+
   def initialize
     @user_interaction = UserInteraction.new
     @create_classes = @user_interaction.create_classes
   end
 
+  def display_choices
+    puts "\nPlease choose an option by enter a number:"
+    puts ['1 - List all books', '2 - List all people', '3 - Create a person',
+          '4 - Create a book', '5 - Create a rental', '6 - List all rentals for a given person id', '7 - Exit']
+  end
+
   def run
-   puts "\nPlease choose an option by enter a number:"
-   puts ['1 - List all books', '2 - List all people', '3 - Create a person',
-        '4 - Create a book', '5 - Create a rental', '6 - List all rentals for a given person id',
-        '7 - Exit']
-   decision = gets.chomp
-   
-   puts 'please choose of the list' unless '1234567'.include?(decision)
-
-   methods = [
-     method(:display_books), method(:display_people), method(:create_people),
-     method(:create_book), method(:handle_rental), method(:list_rental_for_person),
-     method(:exit_program)
-   ]
-
-  '1234567'.include?(decision) && methods[decision.to_i - 1].call     
+    decision = gets.chomp
+    puts 'please choose of the list' unless '1234567'.include?(decision)
+    methods = [
+      method(:display_books), method(:display_people), method(:create_people), method(:create_book),
+      method(:handle_rental), method(:list_rental_for_person), method(:exit_program)
+    ]
+    '1234567'.include?(decision) && methods[decision.to_i - 1].call
   end
 
   def display_books
@@ -31,21 +30,20 @@ class App
   end
 
   def display_people
-     Display.new.display_people(@create_classes.people_list)
+    Display.new.display_people(@create_classes.people_list)
   end
 
   def create_people
-      @user_interaction.create_people
+    @user_interaction.create_people
   end
 
   def create_book
-     @user_interaction.create_book
+    @user_interaction.create_book
   end
 
   def handle_rental
-      @user_interaction.handle_rental
+    @user_interaction.handle_rental
   end
-
 
   def list_rental_for_person
     Display.new.display_rentals(@create_classes.rental_list)
@@ -55,5 +53,4 @@ class App
     puts 'Thank you for using this app!'
     exit
   end
-
 end
