+++ main.rb
@@ -4,6 +4,7 @@ def main
   puts "\nWelcome to School Library App!\n"
   app = App.new
   loop do
+    app.display_choices
     app.run
   end
 end
